+++ src/util.ts
@@ -1,6 +1,8 @@
 import _ from 'lodash/fp.js'
 import { type Document } from 'mongodb'
 
+import { BulkQueryResult } from './crate.js'
+
 /**
  * Does arr start with startsWith array.
  */
@@ -22,19 +24,41 @@ export const setDefaults = (keys: string[], val: any) => {
   return obj
 }
 
-export const renameKey = (doc: Document, key: string, newKey: string) =>
-  _.flow(_.set(newKey, _.get(key, doc)), _.omit([key]))(doc)
+export const renameKey = (doc: Document, key: string, newKey: string) => {
+  const temp = doc[key]
+  delete doc[key]
+  doc[newKey] = temp
+}
 
+/**
+ * Rename keys, mutating the given object.
+ */
 export const renameKeys = (doc: Document, keys: Record<string, string>) => {
-  let newDoc = doc
   for (const key in keys) {
-    if (_.has(key, doc)) {
+    if (key in doc) {
       const newKey = keys[key]
-      newDoc = renameKey(newDoc, key, newKey)
+      renameKey(doc, key, newKey)
     }
   }
-  return newDoc
 }
 
 export const sumByRowcount = (num: number) =>
   _.sumBy(({ rowcount }) => (rowcount === num ? 1 : 0))
+
+/**
+ * Get the documents that failed to be written during a bulk
+ * query.
+ */
+export const getFailedRecords = (
+  results: BulkQueryResult['results'],
+  documents: Document[]
+) => {
+  const failed: unknown[] = []
+  for (let i = 0; i < results.length; i++) {
+    const result = results[i]
+    if (result.rowcount === -2) {
+      failed.push(documents[i])
+    }
+  }
+  return failed
+}