+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.47.0
+
+-   Breaking Change: Default to dynamic object policy in `convertSchema`. See docs on `strictMode` option.
+-   Added `failedRecords` to `process` event and `failedRecord` to `error` for changestream events where applicable.
+-   Added `options.mapper` to `initSync`.
+
 # 0.46.0
 
 -   Bump packages, including latest `prom-utils` which allows for throttling of items/sec and bytes/sec.